// webui/src/generated/api/model/dtoContestResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoContestResponse {
  beginAt: string;
  createdAt: string;
  createdBy?: string;
  description: string;
  endAt: string;
  freezeAt?: string;
  id: string;
  rankboardVisible: boolean;
  rule: string;
  title: string;
  visibility: string;
}